fix: fixed table rendering (html and gfm) and reorganised mdx plugins
