+++ package.json
@@ -60,6 +60,7 @@
     "husky": "8.0.3",
     "lint-staged": "14.0.1",
     "mdast-util-gfm-autolink-literal": "~2.0.0",
+    "mdast-util-gfm-table": "~1.0.7",
     "next": "~13.5.3",
     "next-mdx-remote": "~4.4.1",
     "next-themes": "~0.2.1",
@@ -72,7 +73,8 @@
     "react-dom": "^18.2.0",
     "react-intl": "~6.4.7",
     "rehype-autolink-headings": "~7.0.0",
-    "rehype-pretty-code": "^0.10.1",
+    "rehype-pretty-code": "~0.10.1",
+    "rehype-raw": "~7.0.0",
     "rehype-slug": "~6.0.0",
     "semver": "~7.5.4",
     "sharp": "0.32.6",
